updated spotprices package, removed old testing code in aws
Spotprices package will be responsible for grabbing the most up to date
spotprices, given a specific profile.

currently is not fully integrated.  will work on that.

These will be stored in the datastore for historical purposes

Premise is prices are returned from a call to getPrice().  This
will query the db or download if necessary.  Lastly, it will spawn
a taskqueue to recheck prices every 5 mins, and then pass new prices as
a message.  (This is so prices are updated automatically to the viewer)
taskqueue
that
cycles
every